Python OOP 2020/OOP_2020_exam_prep/Exam_16.08.2020/project/hardware/hardware.py


from project.software.software import Software


class Hardware:

    def __init__(self, name: str, type: str, capacity: int, memory: int):
        self.name = name
        self.type = type
        self.capacity = capacity
        self.memory = memory
        self.software_components = []

    def install(self, software: Software):
        if software.memory_consumption <= self.memory and software.capacity_consumption <= self.capacity:
            self.software_components.append(software)
        else:
            raise Exception("Software cannot be installed")

    def uninstall(self, software: Software):
        self.software_components.remove(software)


